38번 이슈 해결
그 뭐냐 탭키 누르면 공격 안나가게 해야할것 같음
