+++ Assets/Scripts/Interact/ItemSpawner.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ee61cb0d788bc4adc94020869f2649ab
+guid: 76007ebdd625140fa96bd1dfd1f4d3d2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
