--- Assets/Scripts/Interact/betteryspawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 76007ebdd625140fa96bd1dfd1f4d3d2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
